Addictional refactoring.
