+++ ConsoleApp3/ConsoleApp3/Program.cs
@@ -1,5 +1,7 @@
 ﻿using System;
 using UniversityPET.dll;
+using UniversityPET.dll.Models.Implementations;
+using UniversityPET.dll.Services.Implementations;
 
 namespace UniversityPET
 {
