+++ ConsoleApp3/my_first_classLibrary/Models/Implementations/Person.cs
@@ -1,4 +1,4 @@
-﻿namespace UniversityPET.dll
+﻿namespace UniversityPET.dll.Models.Implementations
 {
     public class Person
     {
