+++ ConsoleApp3/my_first_classLibrary/Models/Implementations/University.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace UniversityPET.dll
+namespace UniversityPET.dll.Models.Implementations
 {
     public class University
     {
