+++ ConsoleApp3/my_first_classLibrary/Services/Implementations/TxtDatabaseService.cs
@@ -1,4 +1,6 @@
-﻿namespace UniversityPET.dll
+﻿using UniversityPET.dll.Services.Interfaces;
+
+namespace UniversityPET.dll.Services.Implementations
 {
     public class TxtDatabaseService<T>
     {
