+++ ConsoleApp3/my_first_classLibrary/Services/Interfaces/IReader.cs
@@ -1,4 +1,4 @@
-﻿namespace UniversityPET.dll
+﻿namespace UniversityPET.dll.Services.Interfaces
 {
     public interface IReader
     {
