+++ ConsoleApp3/my_first_classLibrary/Services/Interfaces/ISerializator.cs
@@ -1,4 +1,4 @@
-﻿namespace UniversityPET.dll
+﻿namespace UniversityPET.dll.Services.Interfaces
 {
     public interface ISerializator<T>
     {
